requirements.txt
flask
python-dotenv
requests
boto3